job-recommender/requirements.txt
fastapi
pydantic[email]
pymongo[srv]
python-dotenv
requests
sentence-transformers
scikit-learn
numpy
bson